requirements.txt
# Core Web Framework
streamlit>=1.28.1
python-dotenv>=1.0.0

# Document Processing
python-docx>=0.8.11
pandas>=2.1.3
numpy>=1.21,<2.0

# AI and ML (with fallback support)
sentence-transformers>=2.7.0
transformers>=4.41.0,<5.0.0

# Optional: OpenAI (enhanced features when API key available)
openai>=1.86.0,<2.0.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0

# Data Visualization
plotly>=5.17.0
matplotlib>=3.8.2

# Utilities
requests>=2.31.0
Pillow>=10.1.0
